revert: "chore: use jack on linux"
This reverts commit 82f469e41f1731d2fd18c08e76bf2424e8e8e354.

this is because jack honestly sucks and alsa memory leak is only like a
few kilobytes, not enough for anyone to actually give a shit. this was
originally found when investigating a memory leak with the (now unused)
minimp3 decoder present in encore (we've now switched to a (rust-based)
decoder called symphonia-mp3 which doesn't leak as much (if not at all))

also, i use artix btw. probably a newer version of alsa had the leak
that has now been fixed probably. who knows. linux audio is a mess.